Integrate MetaCoq's optimization; make extraction pipeline more modular (#68)
Integrate the "prop discrimination" optimization pass from MetaCoq that
removes matches and projections on boxes. At the same time I made our
extraction more modular so it is easier to add and integrate new passes
in the future.

Also fix a bug in Rust extraction where we were not emitting proper
PhantomData's for unused type arguments.